Remover filtros de grupo e tipo e tornar o filtro por equipamento (modelo) trazendo apenas modelos do cliente selecionado

+++ SAT.MODELS/Entities/RelatorioAtendimentoDetalhe.cs
@@ -17,7 +17,7 @@ namespace SAT.MODELS.Entities
         public int? CodOrigemCausa { get; set; }
         public int CodAcao { get; set; }
         public int? CodServico { get; set; }
-        public int? IndDefeitoMaquina { get; set; }
+        public byte? IndDefeitoMaquina { get; set; }
         public string CodUsuarioCad { get; set; }
         public int CodOS { get; set; }
         public DateTime? DataHoraCad { get; set; }
